refactor(partenaires): réorganise la définition des interfaces
