+++ src/domain/commun/withCache.interface.ts
@@ -1,0 +1,10 @@
+/**
+ * Interface for classes that used cache mechanism needed to be refreshed after
+ * some operations.
+ */
+interface WithCache {
+  /**
+   * Refresh the cache from the source data.
+   */
+  loadCache(): Promise<void>;
+}
