+++ src/domain/partenaires/partenaireDefinition.ts
