+++ src/usecase/articles.usecase.ts
@@ -11,7 +11,7 @@ export class ArticlesUsecase {
   ) {}
 
   public async updateAllPartenairesCodes(block_size = 100) {
-    await this.partenaireUsecase.updateCodesFromPartenaireFor(
+    await this.partenaireUsecase.updateAllFromPartenaireCodes(
       this.articleRepository,
       block_size,
     );
