+++ test/unit/domain/aide.spec.ts
@@ -2,7 +2,7 @@ import { Aide } from '../../../src/domain/aides/aide';
 import { AideDefinition } from '../../../src/domain/aides/aideDefinition';
 import { Besoin } from '../../../src/domain/aides/besoin';
 import { Echelle } from '../../../src/domain/aides/echelle';
-import { PartenaireDefinition } from '../../../src/domain/contenu/partenaireDefinition';
+import { PartenaireDefinition } from '../../../src/domain/partenaires/partenaireDefinition';
 
 const aide_def = new AideDefinition({
   content_id: '1',
